refactor: restructure project with clean architecture
